Move getUsersPipelineConfig/getNamespace to jenkins-client-plugin
In order to be able to use vanilla openshift image with no customization we need
to have the pipeline not using io.fabric8.kubernetes.client this moves the
getUsersPipelineConfig/ and getNamespace methods to use instead the
openshift-jenkins-client-plugin :

https://github.com/openshift/jenkins-client-plugin#introduction-to-selectors

Which is shipped into the upstream openshift image.

There is quite a long way to go but this gets the example for how to do that.

There is no unittest in that repo (very sad panda), but this is the way i have
tested that. I spun a jenkins on a minishift (i.e: vanilla) and created a test
repo with this JenkinsFile :

https://paste.fedoraproject.org/paste/ix72yD-Af5DqL2cs9GrEQQ

which ran then fine and was able to pull up the key,

Not ideal, not ideal but "c'est la vie" like they say overhere,